requirements.txt
adafruit-io==2.7.1
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
paho-mqtt==1.6.1
pyserial==3.5
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.0
